+++ mono/test/commands/format_command_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:mono/src/commands/format.dart';
 import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 import 'package:test/test.dart';
 import 'package:path/path.dart' as p;