+++ mono/test/commands/task_command_test.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:mono/src/commands/task.dart';
 import 'package:mono_cli/mono_cli.dart';
 import 'package:test/test.dart';
+import 'package:mono_core/mono_core.dart';
 import 'package:path/path.dart' as p;
 
 import '../util/fs_fixtures.dart';