+++ mono/test/config_io_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:mono_cli/mono_cli.dart';
 import 'package:test/test.dart';
+import 'package:mono_core/mono_core.dart';
 
 void main() {
   group('workspace config filesystem helpers', () {