+++ mono_cli/lib/src/external_exports/mono_core.dart
@@ -1,1 +1,0 @@
-export 'package:mono_core/mono_core.dart';