+++ mono_cli/lib/src/ports/graph_builder_impl.dart
@@ -1,4 +1,4 @@
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 @immutable
 class DefaultGraphBuilder implements GraphBuilder {