+++ mono_cli/lib/src/src.dart
@@ -5,7 +5,6 @@ export 'config/yaml_loader.dart';
 export 'config/yaml_validator.dart';
 export 'external_exports/args.dart';
 export 'external_exports/glob.dart';
-export 'external_exports/mono_core.dart';
 export 'external_exports/path.dart';
 export 'external_exports/pool.dart';
 export 'external_exports/yaml.dart';