+++ mono_cli/lib/src/system_io/logger.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class StdLogger implements Logger {
   const StdLogger();