+++ mono_cli/lib/src/system_io/process_runner.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class DefaultProcessRunner implements ProcessRunner {
   const DefaultProcessRunner();