+++ mono_cli/test/ports/task_executor_impl_test.dart
@@ -1,5 +1,6 @@
 import 'package:mono_cli/mono_cli.dart' hide equals;
 import 'package:test/test.dart';
+import 'package:mono_core/mono_core.dart';
 
 class _BufferingLogger implements Logger {
   _BufferingLogger(this.out, this.err);